Restructure and restore Arcee CLI
