+++ arcee/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.2.2"
+__version__ = "1.3.0"
 
 import os
 
@@ -28,10 +28,8 @@ from arcee.dalm import DALM, DALMFilter
 
 if not config.ARCEE_API_KEY:
     # We check this because it's impossible the user imported arcee, _then_ set the env, then imported again
+    # We don't want to block or prompt here because this will interfere with CLI usage
     config.ARCEE_API_KEY = os.getenv("ARCEE_API_KEY", "")
-    while not config.ARCEE_API_KEY:
-        config.ARCEE_API_KEY = input("ARCEE_API_KEY not found in environment. Please input api key: ")
-    os.environ["ARCEE_API_KEY"] = config.ARCEE_API_KEY
 
 __all__ = [
     "upload_docs",
