+++ arcee/cli/errors.py
@@ -1,0 +1,5 @@
+from click.exceptions import ClickException
+
+
+class ArceeException(ClickException):
+    pass
